Entirety of 0.1.1
First and foremost, I have added the .set_rotor_as() and .rotors() methods to set rotor values and view rotor values, as changing the variables was a bit unprofessional.
I have also added a .quick_plug_board() method to set up the plugboard quickly.

For the bug fixes, I've fixed the bug where the code looked for the key '9fs' instead of '9s' which made the 9th plugboard slot unusable. This has been fixed.
The logic also only added used letters in pairs, which allowed plugboard slots like 'a':'a' to exist, which while not breaking the code, still wasn't intentional, so it has been patched.
Lastly, there was a bug that kept offsetting the rotors by -1 each time the method enigma was run, the reason and story for this I cannot explain here for it is too big. Regardless, it has been fixed.